--- a/pkg/eventfilter/benchmarks/common_benchmark_test.go
+++ b/pkg/eventfilter/benchmarks/common_benchmark_test.go
@@ -43,6 +43,7 @@ func RunFilterBenchmarks(b *testing.B, filterCtor func(interface{}) eventfilter.
 		b.Run("Creation: "+fb.name, func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
 				Filter = filterCtor(fb.arg)
+				Filter.Cleanup()
 			}
 		})
 		// Filter to use for the run