--- a/pkg/eventfilter/subscriptionsapi/all_filter.go
+++ b/pkg/eventfilter/subscriptionsapi/all_filter.go
@@ -18,34 +18,88 @@ package subscriptionsapi
 
 import (
 	"context"
+	"sync"
 
 	cloudevents "github.com/cloudevents/sdk-go/v2"
+	"go.uber.org/atomic"
 	"go.uber.org/zap"
 	"knative.dev/pkg/logging"
 
 	"knative.dev/eventing/pkg/eventfilter"
 )
 
-type allFilter []eventfilter.Filter
+type filterCount struct {
+	filter eventfilter.Filter
+	count  atomic.Uint64
+}
+
+type allFilter struct {
+	filters []filterCount
+	rwMutex sync.RWMutex
+	c       chan int
+	d       chan bool
+}
 
 // NewAllFilter returns an event filter which passes if all the contained filters pass
 func NewAllFilter(filters ...eventfilter.Filter) eventfilter.Filter {
-	return append(allFilter{}, filters...)
+	filterCounts := make([]filterCount, len(filters))
+	for i, filter := range filters {
+		filterCounts[i] = filterCount{
+			count:  *atomic.NewUint64(uint64(0)),
+			filter: filter,
+		}
+	}
+	f := &allFilter{
+		filters: filterCounts,
+		c:       make(chan int),
+		d:       make(chan bool),
+	}
+	go f.optimizeLoop()
+	return f
 }
 
-func (filter allFilter) Filter(ctx context.Context, event cloudevents.Event) eventfilter.FilterResult {
+func (filter *allFilter) Filter(ctx context.Context, event cloudevents.Event) eventfilter.FilterResult {
 	res := eventfilter.NoFilter
 	logging.FromContext(ctx).Debugw("Performing an ALL match ", zap.Any("filters", filter), zap.Any("event", event))
-	for _, f := range filter {
-		res = res.And(f.Filter(ctx, event))
+	filter.rwMutex.RLock()
+	defer filter.rwMutex.RUnlock()
+	for i, f := range filter.filters {
+		res = res.And(f.filter.Filter(ctx, event))
 		// Short circuit to optimize it
 		if res == eventfilter.FailFilter {
+			filter.c <- i
 			return eventfilter.FailFilter
 		}
 	}
 	return res
 }
 
-func (filter allFilter) Cleanup() {}
+func (filter *allFilter) optimizeLoop() {
+	for {
+		i, more := <-filter.c
+		if !more {
+			filter.d <- true
+			return
+		}
+		val := filter.filters[i].count.Inc()
+		if i != 0 && val > filter.filters[i-1].count.Load()*2 {
+			go filter.optimize(i)
+		}
+	}
+}
 
-var _ eventfilter.Filter = allFilter{}
+func (filter *allFilter) optimize(swapIdx int) {
+	filter.rwMutex.Lock()
+	defer filter.rwMutex.Unlock()
+	filter.filters[swapIdx-1], filter.filters[swapIdx] = filter.filters[swapIdx], filter.filters[swapIdx-1]
+}
+
+func (filter *allFilter) Cleanup() {
+	close(filter.c)
+	<-filter.d
+	for _, f := range filter.filters {
+		f.filter.Cleanup()
+	}
+}
+
+var _ eventfilter.Filter = &allFilter{}